refactor(interactive-section): extract useDocumentStepProgress hook [Pattern J]
Phase: Tier B
Step: B3
Pattern: J (contract-surface ownership move)
Risk: medium

Why: The effect that publishes document-wide step progress owns
three external contracts:
  - `window.__DocsPluginTotalSteps` (read by lib/analytics.ts)
  - `window.__DocsPluginCurrentStepIndex` (read by lib/analytics.ts)
  - `pathfinder-step-progress` CustomEvent (consumed by
    src/hooks/useStepProgressFromEvents.ts to drive the
    FullScreenLayout / FloatingPanel progress chip)

Per the Guidelines (Pattern J — "Do not rename or clean up contract
surfaces during structural refactor"), this commit moves ownership
into a focused hook without changing names, payload shapes, or
dispatch timing. The contracts tripwire from Phase 0 is the gate.

The hook is a pure side-effect (no return value). Effect deps are
identical to the pre-extraction effect:
`[currentlyExecutingStep, stepComponents, sectionId, completedSteps]`.

Tests:
  - No new dedicated test file — the contracts tripwire
    (interactive-section.contracts.tripwire.test.tsx) provides
    permanent coverage for the payload shape across the seam. Per
    Principle 4: when an extraction's "pre-test" tripwire is
    structurally equivalent to the post-test coverage, the same
    test serves both roles.

Checkpoint:
  - npm run typecheck: pass
  - npx eslint: pass
  - npx prettier --check: pass
  - npx jest src/components/interactive-tutorial: 19 suites pass,
    294 tests pass + 5 todos + 12 pre-existing skips.
    Contracts tripwire's `pathfinder-step-progress` and
    `__DocsPluginTotalSteps` assertions still pass.
  - interactive-section.tsx line delta: 1641 → 1604 lines (-37).

Notes:
  - The `try/catch` with no-op catch is preserved literally; it
    guards against rare environments where `window.dispatchEvent`
    isn't available. Not part of the contract.
  - `interactiveStepStorage` is still imported by
    interactive-section.tsx for the `countAllCompleted` call inside
    `persistCompletedSteps` (which moves to B5). The import line
    stays until then.
